% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf.R
\name{add_pop_to_VCF}
\alias{add_pop_to_VCF}
\title{Add a Population object to a VCF object}
\usage{
add_pop_to_VCF(vcf, pop)
}
\arguments{
\item{vcf}{A VCF object. This should be an object created using the relevant VCF functions in the package.}

\item{pop}{A Population object. This should be an object created using the relevant Population functions in the package.}
}
\value{
A new VCF object that includes the added Population information.
}
\description{
This function adds a Population object to an existing VCF object and creates a new VCF object
that includes the population information.
}
\examples{
# Assume 'vcf' is a valid VCF object and 'pop' is a valid Population object
new_vcf <- add_pop_to_VCF(vcf, pop)
summary(new_vcf) # Inspect the new VCF object
}
