% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BaseInfo.R
\name{add_trait_A}
\alias{add_trait_A}
\title{Add Trait with Additive Effects to BaseInfo}
\usage{
add_trait_A(
  info,
  name,
  mean,
  h2,
  sd = NULL,
  a = NULL,
  loci = NULL,
  num_loci = 1
)
}
\arguments{
\item{info}{External pointer to BaseInfo object}

\item{name}{Name of the trait}

\item{mean}{Phenotype mean}

\item{h2}{Heritability}

\item{sd}{Optional. Phenotype standard deviation}

\item{a}{Optional. Numeric vector of additive effects}

\item{loci}{Optional. List of loci in the form of a data frame with two columns: 'chrom' and 'marker'. For example:
\preformatted{
chrom <- c(3, 5, 10)
marker <- c(1, 2, 1000)
loci <- data.frame(chrom, marker)
}}

\item{num_loci}{Optional. Number of loci (default is 1)}
}
\description{
This function adds a trait with additive effects to the BaseInfo object.
}
\examples{
# Assume 'info' is a valid BaseInfo object
add_trait_A(info, "Trait1", mean=100.0, h2=0.6, sd=10.0, num_loci = 2)
trait <- get_trait(info, 1)
summary(trait)
}
