% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BaseInfo.R
\name{add_trait_AD}
\alias{add_trait_AD}
\title{Add Trait with Additive and Dominance Effects to BaseInfo}
\usage{
add_trait_AD(
  info,
  name,
  mean,
  sd = NULL,
  h2 = NULL,
  H2 = NULL,
  a = NULL,
  d = NULL,
  loci = NULL,
  num_loci = 1
)
}
\arguments{
\item{info}{External pointer to BaseInfo object}

\item{name}{Name of the trait}

\item{mean}{Phenotype mean}

\item{sd}{Optional. Phenotype standard deviation}

\item{h2}{Optional. Narrow-sense heritability (proportion of variance due to additive genetic effects)}

\item{H2}{Optional. Broad-sense heritability (proportion of variance due to all genetic effects, can be NULL)}

\item{a}{Optional. Numeric vector of additive effects}

\item{d}{Optional. Numeric vector of dominance effects}

\item{loci}{Optional. List of loci in the form of a data frame with two columns: 'chrom' and 'marker'. For example:
\preformatted{
chrom <- c(3, 5, 10)
marker <- c(1, 2, 1000)
loci <- data.frame(chrom, marker)
}}

\item{num_loci}{Optioal. Number of loci (default is 1)}
}
\description{
This function adds a trait with additive and dominance effects to the BaseInfo object.
}
\examples{
# Assume 'info' is a valid BaseInfo object
add_trait_AD(info, "Trait1", mean=100.0, h2=0.6, H2=0.7, sd=10.0, num_loci = 2)
trait <- get_trait(info, 1)
summary(trait)
}
