% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BaseInfo.R
\name{addTraitAD}
\alias{addTraitAD}
\title{Add Trait with Additive and Dominance Effects to BaseInfo}
\usage{
addTraitAD(
  info,
  name,
  mean,
  sd = NULL,
  h2 = NULL,
  H2 = NULL,
  a = NULL,
  d = NULL,
  loci = NULL,
  num_loci = 1
)
}
\arguments{
\item{info}{External pointer to BaseInfo object}

\item{name}{Name of the trait}

\item{mean}{Phenotype mean}

\item{a}{Optional numeric vector of additive effects}

\item{d}{Optional numeric vector of dominance effects}

\item{loci}{Optional list of loci}

\item{num_loci}{Number of loci (default is 1)}

\item{Optional}{H2 Broad-sense heritability (proportion of variance due to all genetic effects, can be NULL)}
}
\description{
This function adds a trait with additive and dominance effects to the BaseInfo object.
}
