% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Population.R
\name{check_parent_existance}
\alias{check_parent_existance}
\title{Check Parent Existence in Population}
\usage{
check_parent_existance(df, mat_pop, pat_pop)
}
\arguments{
\item{df}{A data.frame representing the cross table containing 'mat' (maternal names) and 'pat' (paternal names) columns.}

\item{mat_pop}{An external pointer to the maternal Population object.}

\item{pat_pop}{An external pointer to the paternal Population object.}
}
\value{
This function does not return a value. It outputs messages if any maternal or paternal names
in the cross table are not found in the respective populations.
}
\description{
This function checks whether the maternal and paternal names in the given cross table
are present in the specified maternal and paternal populations.
}
\examples{
# Assuming 'df', 'mat_pop', and 'pat_pop' are valid objects
check_parent_existance(df, mat_pop, pat_pop)
}
