% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BaseInfo.R
\name{create_base_info}
\alias{create_base_info}
\title{Create a BaseInfo object}
\usage{
create_base_info(
  positions = NULL,
  num_markers = 1000,
  bp = 1e+08,
  chrom_maps = NULL,
  num_chroms = 10,
  cM = 100,
  seed = -1
)
}
\arguments{
\item{positions}{Optional. A list of numeric vectors, where each vector represents
the marker positions for a chromosome in base pairs. The number of elements in
\code{positions} defines the number of chromosomes and overrides \code{num_chroms}. If \code{positions}
is provided, \code{num_markers} is ignored.}

\item{num_markers}{Optional. An integer. Number of markers per chromosome. Ignored if
\code{positions} is provided. Default is 1000.}

\item{bp}{Optional. An integer. Length of each chromosome in base pairs. Ignored if
\code{positions} is provided. Default is 100000000.}

\item{chrom_maps}{Optional. A list of data.frames, each representing a chromosome map.
Each data.frame should have two columns: 'cM' for centiMorgans and
'position' for base pair positions. The list should be named, with each name
corresponding to a chromosome identifier (e.g., "chr1", "chr2", etc.).
For a given \code{cM} value, the corresponding \code{position} can be interpolated or extrapolated,
and vice versa. This effectively represents the mapping as a piecewise linear
approximation of the chromosome's relationship between cM and base pair positions.
If \code{chrom_maps} is provided, the parameter \code{num_chroms} is ignored.}

\item{num_chroms}{Optional. An integer. Number of chromosomes. Ignored if either
\code{chrom_maps} or \code{positions} is provided. Default is 10.}

\item{cM}{Optional. A numeric. Length of each chromosome in centiMorgans. Ignored if
\code{chrom_maps} is provided. Default is 100.}

\item{seed}{Optional. An integer. A seed for random number generation. Default is -1,
which generates a random seed.}
}
\value{
An external pointer to a BaseInfo object.
}
\description{
This function creates a BaseInfo object. If the seed is set to -1, a random
seed is generated, resulting in different outcomes each time the function
is called. If a specific seed is provided, the random number generation will
be based on that seed, ensuring reproducible results.
}
\examples{
# Create a BaseInfo object with a random seed
base_info_random <- create_base_info()
get_info(base_info_random)

# Create a BaseInfo object with a specific seed for reproducible results
base_info_reproducible <- create_base_info(seed = 123)
get_info(base_info_reproducible)

# Create marker positions manually
num_chr <- 10
position <- sapply(1:100, function(i) i * 1000000)
positions <- replicate(10, position, simplify = FALSE)

# Create a chromosome map with 100 cM and 100 Mbp
cM <- c(14.0, 27.9, 40.2, 48.3, 50.0, 51.7, 59.8, 72.1, 86.0, 100.0)
chr_position <- sapply(1:10, function(i) i * 10000000)
chrom_map <- data.frame(cM, chr_position)
chrom_maps <- replicate(10, chrom_map, simplify = FALSE)
names(chrom_maps) <- paste0("chr", 1:10)
info <- create_base_info(positions = positions, chrom_maps = chrom_maps, seed = 123)
get_info(info)
}
