% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Population.R
\name{create_pop_from_HaploArray}
\alias{create_pop_from_HaploArray}
\title{Create a Population object from a HaploArray}
\usage{
create_pop_from_HaploArray(haploArray, info)
}
\arguments{
\item{haploArray}{A 3-dimensional array where:
\describe{
\item{dim\link{1}}{Number of individuals.}
\item{dim\link{2}}{Number of markers.}
\item{dim\link{3}}{Size must be 2, representing Maternal and Paternal alleles.}
}
Allele values must be binary (0 or 1), based on the reference genome.}

\item{info}{An external pointer to a BaseInfo object.}
}
\value{
An external pointer to a Population object.
}
\description{
This function takes a 3-dimensional array (HaploArray) and a BaseInfo object to create a Population object.
The HaploArray should have dimensions representing individuals, markers, and alleles (Maternal and Paternal).
Allele values must be binary (0 or 1). A value of 0 indicates that the allele matches the reference genome,
while a value of 1 indicates the alternative allele.
The function validates the input data and initializes the Population object accordingly.
}
\examples{
# Create a HaploArray
haploArray <- array(
  data = rbinom(n = 30, size = 1, prob = 0.3),
  dim = c(3, 5, 2),
  dimnames = list(
    paste0("Ind_", 1:3),
    paste0("Mrk_", 1:5),
    c("Maternal", "Paternal")
  )
)

# Assuming 'info' is a valid BaseInfo object
pop <- create_pop_from_HaploArray(haploArray, info)
summary(pop)
}
