% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BaseInfo.R
\name{createBaseInfo}
\alias{createBaseInfo}
\title{Create a BaseInfo object}
\usage{
createBaseInfo(seed)
}
\arguments{
\item{seed}{An integer. A seed for random number generation.}
}
\value{
An external pointer to a BaseInfo object.
}
\description{
Create a BaseInfo object
}
