% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Population.R
\name{createOrigins}
\alias{createOrigins}
\title{Create origins for a Population object}
\usage{
createOrigins(num_inds, info, name_base)
}
\arguments{
\item{num_inds}{An integer. The number of individuals.}

\item{info}{An external pointer to a BaseInfo object.}

\item{name_base}{A string. The base name for individuals.}
}
\value{
An external pointer to a Population object.
}
\description{
Create origins for a Population object
}
