% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Population.R
\name{cross}
\alias{cross}
\title{Cross two Population}
\usage{
cross(num_inds, mothers, fathers, info, name_base, T)
}
\arguments{
\item{num_inds}{An integer. The number of individuals.}

\item{mothers}{An external pointer to a Population object.}

\item{fathers}{An external pointer to a Population object.}
}
\value{
An external pointer to a Population object.
}
\description{
Cross two Population
}
