% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Population.R
\name{generate_names}
\alias{generate_names}
\title{Generate names for individuals based on a base name and number of individuals}
\usage{
generate_names(name_base, num_inds)
}
\arguments{
\item{name_base}{A character string used as the base for generating names.
For example, \code{name_base = "p"} will generate names like \verb{"p1", "p2", ...}.}

\item{num_inds}{A positive integer indicating the number of names to generate.}
}
\value{
A character vector of generated names.
}
\description{
This function creates a vector of names for individuals when \code{names} is not provided.
It uses a base name (\code{name_base}) and a specified number of individuals (\code{num_inds})
to generate names in the format \code{name_base1}, \code{name_base2}, ..., up to the total number.
}
\examples{
# Generate 3 names with base "p"
generated_names <- generate_names("p", 3)
print(generated_names) # Output: "p1", "p2", "p3"
}
