% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Population.R
\name{get_genetypic_values}
\alias{get_genetypic_values}
\title{Get genetypic values from a Population object}
\usage{
get_genetypic_values(pop, i)
}
\arguments{
\item{pop}{An external pointer to a Population object.}

\item{i}{An integer index representing the trait for which genetypic values
are to be retrieved. The index should be between 1
and the total number of traits available in the Population object.}
}
\value{
A vector of genetypic values.
}
\description{
Get genetypic values from a Population object
}
\examples{
# Assuming 'pop' is a valid Population object and trait index 1 is valid
genetypic_values <- get_genetypic_values(pop, 1)
print(genetypic_values)
}
