% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Population.R
\name{get_genotypes}
\alias{get_genotypes}
\title{Get genotypes from a Population object}
\usage{
get_genotypes(pop)
}
\arguments{
\item{pop}{An external pointer to a Population object.}
}
\value{
A matrix of genotypes where rows are individuals and columns are markers.
}
\description{
This function retrieves the genotypes from a given Population object.
The genotypes are represented in a matrix format where rows correspond to
individuals and columns correspond to markers.
}
\details{
Genotype encoding:\cr
\itemize{
\item 0/0 is encoded as -1\cr
\item 0/1 is encoded as 0\cr
\item 1/1 is encoded as 1
}
}
