% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BaseInfo.R
\name{get_info}
\alias{get_info}
\title{Retrieve key values from a BaseInfo object}
\usage{
get_info(info)
}
\arguments{
\item{info}{An external pointer to a BaseInfo object. This object must be properly initialized
and created using the appropriate BaseInfo constructor or relevant functions.}
}
\value{
A list containing the following elements:
\describe{
\item{num_chroms}{An integer representing the total number of chromosomes contained in the BaseInfo object.}
\item{num_traits}{An integer indicating the number of traits managed by the BaseInfo object.}
\item{num_markers}{An integer specifying the total number of markers tracked in the BaseInfo object.}
}
}
\description{
This function extracts essential information from a BaseInfo object. A BaseInfo object
stores metadata related to genetic simulations, such as the number of chromosomes, traits,
and markers. This function is useful for summarizing or verifying the content of a BaseInfo
object during analysis.
}
\examples{
# Assume 'info' is a valid BaseInfo object
values <- get_info(info)
print(values$num_chroms)  # Print the number of chromosomes
print(values$num_traits)  # Print the number of traits
print(values$num_markers) # Print the number of markers
}
