% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Population.R
\name{get_phased_genotypes}
\alias{get_phased_genotypes}
\title{Get phased genotypes from a Population object}
\usage{
get_phased_genotypes(pop)
}
\arguments{
\item{pop}{An external pointer to a Population object.}
}
\value{
A matrix of genotypes where rows are individuals and columns are markers.
}
\description{
This function retrieves the genotypes from a given Population object.
The genotypes are represented in a matrix format where rows correspond to markers
and columns correspond to individuals.
}
\details{
Genotype is 0|0, 0|1, 1|0, or 1|1
}
\examples{
info <- create_base_info()
pop <- create_origins(info, num_inds = 2, names = c("p1", "p2"))
geno <- get_phased_genotypes(pop)
geno[, 1:5]
}
