% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Population.R
\name{get_sample_names}
\alias{get_sample_names}
\title{Get name data from a Population object}
\usage{
get_sample_names(pop)
}
\arguments{
\item{pop}{An external pointer to a Population object.}
}
\value{
A data.frame containing the following columns:
\describe{
\item{name}{Names from the Population object}
\item{mat}{Maternal names from the Population object}
\item{pat}{Paternal names from the Population object}
}
}
\description{
Get name data from a Population object
}
\examples{
# Assuming 'pop' is a valid Population object
name_data <- get_sample_names(pop)
print(name_data)
}
