% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BaseInfo.R
\name{get_trait}
\alias{get_trait}
\title{Get a trait from a BaseInfo object}
\usage{
get_trait(info, i)
}
\arguments{
\item{info}{An external pointer to a BaseInfo object.}

\item{i}{An integer. The index of the trait to retrieve.}
}
\value{
The trait at the specified index.
}
\description{
Get a trait from a BaseInfo object
}
