% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf.R
\name{get_VCF_info}
\alias{get_VCF_info}
\title{Get information from a VCF object}
\usage{
get_VCF_info(vcf)
}
\arguments{
\item{vcf}{A VCF object. This should be an object created using the relevant VCF functions in the package.}
}
\value{
A list containing:
- \code{num_individuals}: The number of individuals in the VCF object.
- \code{num_markers}: The number of markers in the VCF object.
}
\description{
This function retrieves detailed information about a VCF object, including the number
of individuals and the number of markers contained within the VCF.
}
\examples{
# Assume 'vcf' is a valid VCF object
vcf_info <- get_VCF_info(vcf)
print(vcf_info) # Displays the number of individuals and markers
}
