% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BaseInfo.R
\name{getInfo}
\alias{getInfo}
\title{Get the values of a BaseInfo object}
\usage{
getInfo(info)
}
\arguments{
\item{info}{An external pointer to a BaseInfo object}
}
\value{
The list of the values of a BaseInfo object
}
\description{
Get the values of a BaseInfo object
}
