% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BaseInfo.R
\name{getMap}
\alias{getMap}
\title{Retrieve Genetic Map Information}
\usage{
getMap(info)
}
\arguments{
\item{info}{An object of class \code{BaseInfo}. This object contains the genetic map information.}
}
\value{
A list of data frames, each representing a chromosome. Each data frame contains two columns:
\itemize{
\item \code{cM}: The centiMorgan positions of the markers.
\item \code{position}: The base pair positions of the markers.
}
}
\description{
This function retrieves the genetic map information from a \code{BaseInfo} object.
}
\examples{
\dontrun{
  # Assuming `info` is a valid BaseInfo object
  map <- getMap(info)
  print(map)
}
}
