% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Population.R
\name{getPhasedGenotypes}
\alias{getPhasedGenotypes}
\title{Get phased genotypes from a Population object}
\usage{
getPhasedGenotypes(pop)
}
\arguments{
\item{pop}{An external pointer to a Population object.}
}
\value{
A matrix of genotypes where rows are samples and columns are markers.
}
\description{
This function retrieves the genotypes from a given Population object.
The genotypes are represented in a matrix format where rows correspond to markers
and columns correspond to samples.
}
\details{
Genotype is 0|0, 0|1, 1|0, or 1|1
}
