% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Population.R
\name{getPhasedIntGenotypes}
\alias{getPhasedIntGenotypes}
\title{Get phased integer genotypes from a Population object}
\usage{
getPhasedIntGenotypes(pop)
}
\arguments{
\item{population}{An external pointer to a Population object.}
}
\value{
A matrix of genotypes where rows are samples and columns are markers.
        Each sample has two rows: the first row is the maternal allele and the
        second row is the paternal allele.
}
\description{
This function retrieves the genotypes from a given Population object.
The genotypes are represented in a matrix format where rows correspond to samples
and columns correspond to markers. Each sample has two rows: the first row
represents the maternal allele and the second row represents the paternal allele.
}
\details{
Genotype is represented as integers: 0 or 1
}
