% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Population.R
\name{getPhenotypes}
\alias{getPhenotypes}
\title{Get phenotypes from a Population object}
\usage{
getPhenotypes(pop, i)
}
\arguments{
\item{i}{An integer index representing the trait for which phenotypes are to
be retrieved. The index should be between 1
and the total number of traits available in the Population object.}

\item{population}{An external pointer to a Population object.}
}
\value{
A vector of phenotypes.
}
\description{
Get phenotypes from a Population object
}
\examples{
# Assuming 'pop' is a valid Population object and trait index 1 is valid
phenotypes <- getPhenotypes(pop, 1)
print(phenotypes)
}
