% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Population.R
\name{getPopInfo}
\alias{getPopInfo}
\title{Get information for a Population object}
\usage{
getPopInfo(ptr)
}
\arguments{
\item{num_inds}{An integer. The number of individuals.}

\item{info}{An external pointer to a BaseInfo object.}

\item{name_base}{A string. The base name for individuals.}
}
\value{
An external pointer to a Population object.
}
\description{
Get information for a Population object
}
