% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BaseInfo.R
\name{getTrait}
\alias{getTrait}
\title{Get a trait from a BaseInfo object}
\usage{
getTrait(baseInfo, i)
}
\arguments{
\item{baseInfo}{An external pointer to a BaseInfo object.}

\item{i}{An integer. The index of the trait to retrieve.}
}
\value{
The trait at the specified index.
}
\description{
Get a trait from a BaseInfo object
}
