% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Population.R
\name{joinPops}
\alias{joinPops}
\title{Join multiple Population objects}
\usage{
joinPops(...)
}
\arguments{
\item{...}{External pointers to Population objects.}
}
\value{
An external pointer to a new Population object
containing the combined individuals.
}
\description{
Join multiple Population objects
}
\examples{
# Assuming 'pop1', 'pop2', and 'pop3' are valid Population objects
combined_pop <- joinPops(pop1, pop2, pop3)
print(combined_pop)
}
