% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Population.R
\name{remove_pop}
\alias{remove_pop}
\title{Remove individuals from a Population object}
\usage{
remove_pop(pop, samples)
}
\arguments{
\item{pop}{An external pointer to a Population object.}

\item{samples}{A vector of either integer indices or character names
representing the individuals to be removed.}
}
\value{
An external pointer to a new Population object
with the specified individuals removed.
}
\description{
Remove individuals from a Population object
}
\examples{
# Assuming 'pop' is a valid Population object
# Remove individuals by index
new_pop <- remove_pop(pop, c(1, 2, 3))
get_individual_names(new_pop)

# Remove individuals by name
new_pop2 <- remove_pop(pop, c("ind1", "ind2"))
get_individual_names(new_pop2)
}
