% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Population.R
\name{select_pop}
\alias{select_pop}
\title{Select individuals from a Population object}
\usage{
select_pop(pop, indices)
}
\arguments{
\item{pop}{An external pointer to a Population object.}

\item{indices}{A vector of integer indices representing the individuals
to be selected.}
}
\value{
An external pointer to a new Population object
containing the selected individuals.
}
\description{
Select individuals from a Population object
}
\examples{
# Assuming 'pop' is a valid Population object and indices are valid
selected_pop <- select_pop(pop, c(1, 2, 3))
get_individual_names(selected_pop)
}
