% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Population.R
\name{selectPop}
\alias{selectPop}
\title{Select individuals from a Population object}
\usage{
selectPop(population, indices)
}
\arguments{
\item{population}{An external pointer to a Population object.}

\item{indices}{A vector of indices to select.}
}
\value{
An external pointer to a new Population object containing the selected individuals.
}
\description{
Select individuals from a Population object
}
