% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Population.R
\name{set_individual_names}
\alias{set_individual_names}
\title{Set individual names for a Population object}
\usage{
set_individual_names(names, pop)
}
\arguments{
\item{names}{A character vector containing the names of individuals to assign.}

\item{pop}{A Population object where the individual names will be set.}
}
\value{
None. The function modifies the Population object directly.
}
\description{
This function assigns names to individuals in a Population object.
The \code{names} argument must be a character vector, and \code{pop} must be an object
of the "Population" class.
}
\examples{
# Create a Population object and set individual names
info <- create_base_info()
pop <- create_origins(2, info, "ind")
set_individual_names(c("sample1", "sample2"), pop)

# Access the updated individual names
get_individual_names(pop)
}
