% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Population.R
\name{set_sample_names}
\alias{set_sample_names}
\title{Set sample names for a Population object}
\usage{
set_sample_names(names, pop)
}
\arguments{
\item{names}{A character vector containing the names of samples to assign.}

\item{pop}{A Population object where the sample names will be set.}
}
\value{
None. The function modifies the Population object directly.
}
\description{
This function assigns names to samples in a Population object.
The \code{names} argument must be a character vector, and \code{pop} must be an object
of the "Population" class.
}
\examples{
# Create a Population object and set sample names
info <- create_base_info()
pop <- create_origins(2, info, "ind")
set_sample_names(c("sample1", "sample2"), pop)

# Access the updated sample names
get_sample_names(pop)
}
