% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BaseInfo.R
\name{setTrait}
\alias{setTrait}
\title{Set a default trait on a BaseInfo object}
\usage{
setTrait(ptr)
}
\arguments{
\item{ptr}{An external pointer to a BaseInfo object}
}
\value{
No return value, called for side effects.
}
\description{
Set a default trait on a BaseInfo object
}
