% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BaseInfo.R
\name{setTraitADMulti}
\alias{setTraitADMulti}
\title{Add a trait with additive and dominance effects to an object}
\usage{
setTraitADMulti(ptr, num_loci, h2, H2)
}
\arguments{
\item{ptr}{An external pointer to a BaseInfo object}

\item{num_loci}{An integer representing the number of loci}

\item{h2}{a double representing a narrow-sense heritability.}

\item{H2}{a double representing a broad-sense heritability.}
}
\value{
No return value, called for side effects.
}
\description{
Add a trait with additive and dominance effects to an object
}
