% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf.R
\name{write_VCF}
\alias{write_VCF}
\title{Write VCF or Population object to a VCF file}
\usage{
write_VCF(obj, filename)
}
\arguments{
\item{obj}{An object of class VCF or Population. The data to write to the VCF file.}

\item{filename}{A string specifying the path to the output VCF file.}
}
\description{
This function writes either a VCF object or a Population object to a VCF file. If the input is a
VCF object, it directly writes the VCF data to the specified file. If the input is a Population
object, it generates VCF data from the Population and writes it to the file.
}
\examples{
# Example 1: Write a Population object to a VCF file
pop <- create_origins(info, num_inds = 2, names = c("p1", "p2"))
write_VCF(pop, "population_output.vcf")

# Example 2: Write a VCF object to a VCF file
vcf <- read_VCF("input.vcf")
write_VCF(vcf, "vcf_output.vcf")
}
