% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BaseInfo.R
\name{addTraitA}
\alias{addTraitA}
\title{Add Trait with Additive Effects to BaseInfo}
\usage{
addTraitA(info, name, mean, h2, sd = NULL, a = NULL, loci = NULL, num_loci = 1)
}
\arguments{
\item{info}{External pointer to BaseInfo object}

\item{name}{Name of the trait}

\item{mean}{Phenotype mean}

\item{h2}{Heritability}

\item{sd}{Optional Phenotype standard deviation}

\item{a}{Optional numeric vector of additive effects}

\item{loci}{Optional list of loci}

\item{num_loci}{Number of loci (default is 1)}
}
\description{
This function adds a trait with additive effects to the BaseInfo object.
}
