% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Population.R
\name{create_origins}
\alias{create_origins}
\title{Create origins for a Population object}
\usage{
create_origins(info, names = NULL, num_inds = NULL, name_base = NULL)
}
\arguments{
\item{info}{An external pointer to a BaseInfo object.}

\item{num_inds}{An integer. The number of individuals.}

\item{name_base}{A string. The base name for individuals.}
}
\value{
An external pointer to a Population object.
}
\description{
Create origins for a Population object
}
\examples{
# Example 1: Specify individual names directly
pop <- create_origins(info, names = c("p1", "p2"))
get_individual_names(pop)

# Example 2: Generate individual names using name_base and num_inds
pop2 <- create_origins(info, num_inds = 2, name_base = "q")
get_individual_names(pop2)
}
