% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Population.R
\name{cross}
\alias{cross}
\title{Cross two Population}
\usage{
cross(num_inds, mothers, fathers, name_base, num_threads = 0)
}
\arguments{
\item{num_inds}{An integer. The number of individuals.}

\item{mothers}{An external pointer to a Population object.}

\item{fathers}{An external pointer to a Population object.}

\item{name_base}{A string. The base name for individuals.}

\item{num_threads}{optional An integer. The number of threads to be used.
If not specified, the function will use
the maximum number of available threads.}
}
\value{
An external pointer to a Population object.
}
\description{
Cross two Population
}
\examples{
# Assuming 'mothers' and 'fathers' are valid Population objects
new_population <- cross(100, mothers, fathers, "prog_")
summary(new_population)
}
