% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Population.R
\name{cross_randomly}
\alias{cross_randomly}
\title{Cross two Population randomly}
\usage{
cross_randomly(num_inds, mat_pop, pat_pop, name_base, num_threads = 0)
}
\arguments{
\item{num_inds}{An integer. The number of individuals.}

\item{mat_pop}{An external pointer to a Population object representing mothers.}

\item{pat_pop}{An external pointer to a Population object representing fathers.}

\item{name_base}{A string. The base name for individuals.}

\item{num_threads}{Optional. An integer. The number of threads to be used.
If not specified, the function will use
the maximum number of available threads.}
}
\value{
An external pointer to a Population object.
}
\description{
Cross two Population randomly
}
\examples{
# Assuming 'mothers' and 'fathers' are valid Population objects
new_population <- cross_randomly(100, mothers, fathers, "prog_")
summary(new_population)
}
