% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Population.R
\name{getGenotypes}
\alias{getGenotypes}
\title{Get genotypes from a Population object}
\usage{
getGenotypes(pop)
}
\arguments{
\item{pop}{An external pointer to a Population object.}
}
\value{
A matrix of genotypes where rows are samples and columns are markers.
}
\description{
This function retrieves the genotypes from a given Population object.
The genotypes are represented in a matrix format where rows correspond to samples
and columns correspond to markers.
}
\details{
Genotype encoding:\cr
- 0/0 is encoded as -1\cr
- 0/1 is encoded as 0\cr
- 1/1 is encoded as 1
}
