% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Population.R
\name{getPhenotypes}
\alias{getPhenotypes}
\title{Get phenotypes from a Population object}
\usage{
getPhenotypes(population, i)
}
\arguments{
\item{population}{An external pointer to a Population object.}

\item{i}{An integer index.}
}
\value{
A vector of phenotypes.
}
\description{
Get phenotypes from a Population object
}
