% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Population.R
\name{getPopInfo}
\alias{getPopInfo}
\title{Get information for a Population object}
\usage{
getPopInfo(pop)
}
\arguments{
\item{pop}{An external pointer to a BaseInfo object.}
}
\value{
A list containing the number of individuals
and the number of chromosomes in the population.
}
\description{
Get information for a Population object
}
