% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Population.R
\name{write_VCF}
\alias{write_VCF}
\title{Write Population to VCF file}
\usage{
write_VCF(pop, filename)
}
\arguments{
\item{pop}{An external pointer to a Population object.}

\item{filename}{A string specifying the path to the output VCF file.}
}
\description{
This function writes a Population object to a VCF file.
}
\examples{
# Assuming 'pop' is a valid Population object
write_VCF(pop, "output.vcf")
}
